refactor code cells out of parser.py and clean up rendering logic
